+++ sample/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/ui/test/ImageSourceTestScreen.common.kt
@@ -145,7 +145,7 @@ class ImageSourceTestScreen : BaseScreen() {
                 } else {
                     null
                 }
-                zoomState.setImage(imageSource, imageInfo)
+                zoomState.setSubsamplingImage(imageSource, imageInfo)
             }
 
             val imagePainter1 = imagePainter