+++ zoomimage-compose-coil2-core/src/main/kotlin/com/github/panpf/zoomimage/CoilZoomAsyncImage.kt
@@ -298,13 +298,13 @@ private fun onState(
                 }
             }
             if (generateResult is SubsamplingImageGenerateResult.Success) {
-                zoomState.setImage(generateResult.subsamplingImage)
+                zoomState.setSubsamplingImage(generateResult.subsamplingImage)
             } else {
-                zoomState.setImage(null as SubsamplingImage?)
+                zoomState.setSubsamplingImage(null as SubsamplingImage?)
             }
         }
     } else {
-        zoomState.setImage(null as SubsamplingImage?)
+        zoomState.setSubsamplingImage(null as SubsamplingImage?)
     }
 }