+++ zoomimage-compose-glide/src/main/kotlin/com/github/panpf/zoomimage/GlideZoomAsyncImage.kt
@@ -250,13 +250,13 @@ private class ResetListener(
                     logger.d { "GlideZoomAsyncImage. ${generateResult.message}. model='$model'" }
                 }
                 if (generateResult is SubsamplingImageGenerateResult.Success) {
-                    zoomState.setImage(generateResult.subsamplingImage)
+                    zoomState.setSubsamplingImage(generateResult.subsamplingImage)
                 } else {
-                    zoomState.setImage(null as SubsamplingImage?)
+                    zoomState.setSubsamplingImage(null as SubsamplingImage?)
                 }
             }
         } else {
-            zoomState.setImage(null as SubsamplingImage?)
+            zoomState.setSubsamplingImage(null as SubsamplingImage?)
         }
     }
 }