+++ zoomimage-compose-sketch4-core/src/commonMain/kotlin/com/github/panpf/zoomimage/SketchZoomAsyncImage.kt
@@ -249,12 +249,12 @@ private fun onState(
             }
         }
         if (generateResult is SubsamplingImageGenerateResult.Success) {
-            zoomState.setImage(generateResult.subsamplingImage)
+            zoomState.setSubsamplingImage(generateResult.subsamplingImage)
         } else {
-            zoomState.setImage(null as SubsamplingImage?)
+            zoomState.setSubsamplingImage(null as SubsamplingImage?)
         }
     } else {
-        zoomState.setImage(null as SubsamplingImage?)
+        zoomState.setSubsamplingImage(null as SubsamplingImage?)
     }
 }