+++ zoomimage-view-coil2-core/src/main/kotlin/com/github/panpf/zoomimage/CoilZoomImageView.kt
@@ -114,13 +114,13 @@ open class CoilZoomImageView @JvmOverloads constructor(
                         }
                     }
                     if (generateResult is SubsamplingImageGenerateResult.Success) {
-                        setImage(generateResult.subsamplingImage)
+                        setSubsamplingImage(generateResult.subsamplingImage)
                     } else {
-                        setImage(null as SubsamplingImage?)
+                        setSubsamplingImage(null as SubsamplingImage?)
                     }
                 }
             } else {
-                setImage(null as SubsamplingImage?)
+                setSubsamplingImage(null as SubsamplingImage?)
             }
         }
     }