+++ zoomimage-view-glide/src/main/kotlin/com/github/panpf/zoomimage/GlideZoomImageView.kt
@@ -115,13 +115,13 @@ open class GlideZoomImageView @JvmOverloads constructor(
                         }
                     }
                     if (generateResult is SubsamplingImageGenerateResult.Success) {
-                        setImage(generateResult.subsamplingImage)
+                        setSubsamplingImage(generateResult.subsamplingImage)
                     } else {
-                        setImage(null as SubsamplingImage?)
+                        setSubsamplingImage(null as SubsamplingImage?)
                     }
                 }
             } else {
-                setImage(null as SubsamplingImage?)
+                setSubsamplingImage(null as SubsamplingImage?)
             }
         }
     }