+++ zoomimage-view-sketch3-core/src/main/kotlin/com/github/panpf/zoomimage/SketchZoomImageView.kt
@@ -111,12 +111,12 @@ open class SketchZoomImageView @JvmOverloads constructor(
                     }
                 }
                 if (generateResult is SubsamplingImageGenerateResult.Success) {
-                    setImage(generateResult.subsamplingImage)
+                    setSubsamplingImage(generateResult.subsamplingImage)
                 } else {
-                    setImage(null as SubsamplingImage?)
+                    setSubsamplingImage(null as SubsamplingImage?)
                 }
             } else {
-                setImage(null as SubsamplingImage?)
+                setSubsamplingImage(null as SubsamplingImage?)
             }
         }
     }